Add import/export feature, add generate PES 2019 DpFileList.bin, decouple messages and fix bugs

+++ PUMM/PUMM/ViewModel/LibraryViewModel.cs
@@ -5,7 +5,7 @@ using System.IO;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-using System.Windows;
+using System.Windows.Forms;
 using System.Windows.Media;
 using System.Windows.Media.Imaging;
 using PUMM.Model;
@@ -24,8 +24,7 @@ namespace PUMM.ViewModel
             this.db = db;
             Modpacks = db.retrieveModpacks(main.Version);
             this.main = main;
-
-            //Print = new MyICommand<string>(printForDebug);
+            
             AdjustItems = new MyICommand<string>(adjustItemsWidth);
             SetAsActive = new MyICommand<Modpack>(setActive);
             DeleteModpack = new MyICommand<Modpack>(deleteModpack);
@@ -44,36 +43,41 @@ namespace PUMM.ViewModel
             get { return modpacks; }
             set {  SetProperty(ref modpacks, value); }
         }
-
-        //public MyICommand<string> Print { get; set; }
+        
         public MyICommand<string> AdjustItems { get; set; }
         public MyICommand<Modpack> SetAsActive { get; set; }
         public MyICommand<Modpack> DeleteModpack { get; set; }
 
-        /*
-        private void printForDebug(string s)
-        {
-            MessageBox.Show(System.AppDomain.CurrentDomain.BaseDirectory);
-        }
-        */
-
         private void adjustItemsWidth(string s)
         {
             if(PanelWidth < 900)
-                ItemWidth = (PanelWidth - 14 * 3 - SystemParameters.VerticalScrollBarWidth) / 3;
+                ItemWidth = (PanelWidth - 14 * 3 - System.Windows.SystemParameters.VerticalScrollBarWidth) / 3;
             else
-                ItemWidth = (PanelWidth - 14 * 4 - SystemParameters.VerticalScrollBarWidth) / 4;
+                ItemWidth = (PanelWidth - 14 * 4 - System.Windows.SystemParameters.VerticalScrollBarWidth) / 4;
         }
 
         private void setActive(Modpack modpack)
         {
+            // Updates current active modpack and its mods
             main.Active = modpack;
+            // Updates active modpack list of mods and sorts it
+            main.Active.Mods = db.getMods(main.Active);
+            // Updates name for editing
+            main.PotentialName = modpack.Name;
         }
 
         private void deleteModpack(Modpack modpack)
         {
-            Modpacks.Remove(modpack);
-            db.deleteModpack(modpack.Id);
+            string[] question = Messages.question("DeleteModpackConfirmation", new string[] { modpack.Name });
+            DialogResult dialogResult = MessageBox.Show(question[0], question[1], MessageBoxButtons.YesNo, MessageBoxIcon.Question);
+            if (dialogResult == DialogResult.Yes)
+            {
+                if (main.Active != null && main.Active.Id == modpack.Id)
+                    main.Active = null;
+
+                Modpacks.Remove(modpack);
+                db.deleteModpack(modpack.Id);
+            }
         }
 
     }
